send real alarm time, first basic working version (without retry logic nor visual feedback for errors)

+++ app/src/main/java/de/l3d00m/ha_updater/HomeassistantPOJO.kt
@@ -12,7 +12,7 @@ object HomeassistantPOJO {
 
     data class Entity(
         @SerializedName("state")
-        val newState: Int
+        val newState: Long
     )
 
 }
